refactor(database): consolidate database helpers and update dependencies
- Merge pagination.helper.js into database.helper.js for better organization
- Rename utilities.helper.js to database.helper.js with improved documentation
- Add charset and collate options to Sequelize configuration
- Update mysql2 dependency to latest version
- Fix import paths in service files to use new consolidated helper
- Add proper cleanup in unit tests with sequelize.close()
- Create new accounts service for user management
- Improve code consistency and remove redundant comments
