+++ package.json
@@ -98,7 +98,7 @@
     "moment": "^2.30.1",
     "moment-timezone": "^0.5.46",
     "morgan": "^1.10.0",
-    "mysql2": "^3.11.4",
+    "mysql2": "^3.15.1",
     "nodemailer": "^6.10.1",
     "rate-limiter-flexible": "^7.2.0",
     "redis": "^4.7.0",
